Burial Records Put API Completed

+++ database/migrations/2024_04_02_141150_create_services_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('ServiceName'); // Adding varchar column for section code
             $table->string('Description'); // Adding varchar column for section code
-            $table->unsignedInteger('ProviderId'); // Adding int column for total graves
+            $table->unsignedBigInteger('ProviderId'); // Adding int column for total graves
 
             $table->foreign('ProviderId')->references('id')->on('service_providers')->onDelete('cascade');
